pyproject.toml
[tool.poetry]
name = "pahome"
version = "0.1.0"
description = ""
authors = ["rtplv <rtplv@yandex.ru>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.75.0"
uvicorn = {extras = ["standard"], version = "^0.17.6"}
pytest = "^7.0.1"
alembic = "^1.7.6"
gunicorn = "^20.1.0"
python-dotenv = "^0.19.2"
psycopg2-binary = "^2.9.3"
pytest-asyncio = "^0.18.2"
gmqtt = "^0.6.11"
loguru = "^0.6.0"
asyncpg = "^0.25.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
